Move build_hl7_message to V2MessageBuilder module
